Moved lib/recaptcha.rb to config/initializers/recaptcha.rb
