+++ config/environment.rb
@@ -24,7 +24,6 @@ require 'sso_cookie_jar'
 require 'set_gdpr_data'
 require 'date_time'
 require 'educator_signup_flow_decorator'
-require 'recaptcha'
 
 # Initialize the Rails application.
 Rails.application.initialize!
